+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qrgrid/react
 
+## 1.3.1
+
+### Patch Changes
+
+- Fix on generate event call
+
 ## 1.3.0
 
 ### Minor Changes